refactor: Create Views folder to gather views
