--- src/components/Trip/index.js
@@ -1,3 +1,0 @@
-import TripView from './TripView'
-
-export default TripView
